==> src/notebooks/02_web_scraping.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import requests

response = requests.get(url="https://en.wikipedia.org/wiki/Community_areas_in_Chicago")

soup = BeautifulSoup(response.content, "html.parser")

title =  soup.find(id = "firstHeading")
print(title.string)

print(soup.prettify())

soup.find_all("table")

for table in soup.find_all("table"):
    print(table.get("class"))

table = soup.find("table", class_ ="wikitable sortable plainrowheaders mw-datatable")
table

for row in table.tbody.find_all("tr")[2: -1]:
    print (row)
    print ("-" * 100)

for row in table.tbody.find_all("tr")[2: -1]:
    cells = row.find_all("td")
    print (cells)

for row in table.tbody.find_all("tr")[2: -1]:
    cells = row.find_all("td")
    print (cells[0].get_text(strip = True))

data = []

for row in table.tbody.find_all("tr")[2: -1]:
    cells = row.find_all("td")
    header_cell = row.find("th")

    area_code = cells[0].get_text(strip = True)
    community_name = header_cell.a.get_text(strip = True)
    #print (area_code, community_name)

    data.append ({"area_code": area_code, "community_name": community_name})

data

community_areas = pd.DataFrame(data)

community_areas["area_code"] = community_areas["area_code"].astype("int")

community_areas 

community_areas.info()

community_areas.to_csv("./../../data/csv/community_areas_master.csv", index = False)
